fix: fetch hugo dep from npm
Use a local hugo binary rather than hoping the user has the right version installed.
Hugo's api and file layout expectations has changed a lot since this site was built.
We should update it at some point, but just building with the version of hugo that
the site was made for is sufficient for now.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>